# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Leshukov Library API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VocabEntry {
  word: string;
  meaning: string;
}

export interface Book {
  id: string;
  title: string;
  author: string;
  lang: string;
  genre: string;
  year?: number | null;
  rating?: number | null;
  synopsis?: string | null;
  quotes?: string | null;
  thoughts?: string | null;
  vocab: VocabEntry[];
  cover?: string | null;
  createdAt: string;
}

export interface CreateBookInput {
  title: string;
  author: string;
  lang: string;
  genre: string;
  year?: number | null;
  rating?: number | null;
  synopsis?: string | null;
  quotes?: string | null;
  thoughts?: string | null;
  vocab: VocabEntry[];
  cover?: string | null;
}

export interface Note {
  id: string;
  bookId: string;
  text: string;
  createdAt: string;
}

export interface CreateNoteInput {
  text: string;
}

export interface AudioPart {
  id: string;
  bookId: string;
  title: string;
  url: string;
  position: number;
  createdAt?: string | null;
}

export interface CreateAudioPartInput {
  title: string;
  url: string;
  position?: number;
}

export type ListBooksParams = {
  lang?: string;
  search?: string;
};