# lib/api-zod/src/generated/types/audioPart.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Leshukov Library API
 * OpenAPI spec version: 0.1.0
 */

export interface AudioPart {
  id: string;
  bookId: string;
  title: string;
  url: string;
  position: number;
  createdAt?: string | null;
}